Revert "515 compatibility update (#64)"
This reverts commit 352323243eccdbcde66dfca6530f77e404ea9137.

+++ code/game/objects/items/stacks/stack.dm
@@ -42,12 +42,11 @@
 	for(var/i in 1 to grind_results.len) //This should only call if it's ground, so no need to check if grind_results exists
 		grind_results[grind_results[i]] *= get_amount() //Gets the key at position i, then the reagent amount of that key, then multiplies it by stack size
 
-/obj/item/stack/grind_requirements(obj/machinery/reagentgrinder/R, silent)
+/obj/item/stack/grind_requirements()
 	if(is_cyborg)
-		if(!silent)
-			to_chat(usr, span_danger("[src] is electronically synthesized in your chassis and can't be ground up!"))
-		return GRIND_IS_CYBORG
-	return FALSE
+		to_chat(usr, "<span class='danger'>[src] is electronically synthesized in your chassis and can't be ground up!</span>")
+		return
+	return TRUE
 
 /obj/item/stack/Initialize(mapload, new_amount, merge = TRUE)
 	if(new_amount != null)
@@ -64,18 +63,14 @@
 			mats_per_unit[SSmaterials.GetMaterialRef(i)] = in_process_mat_list[i]
 			custom_materials[i] *= amount
 	. = ..()
-	var/static/list/loc_connections = list(
-		COMSIG_ATOM_ENTERED = PROC_REF(on_movable_entered_occupied_turf),
-	)
-	AddElement(/datum/element/connect_loc, loc_connections)
 
 	if(merge)
 		for(var/obj/item/stack/item_stack in loc)
 			if(item_stack == src)
 				continue
 			if(can_merge(item_stack))
-				INVOKE_ASYNC(src, PROC_REF(merge_without_del), item_stack)
-				if(zero_amount())
+				merge_without_del(item_stack)
+				if(is_zero_amount(delete_if_zero = FALSE))
 					return INITIALIZE_HINT_QDEL
 	var/list/temp_recipes = get_main_recipes()
 	recipes = temp_recipes.Copy()
@@ -88,6 +83,10 @@
 					recipes += temp
 	update_weight()
 	update_icon()
+	var/static/list/loc_connections = list(
+		COMSIG_ATOM_ENTERED = PROC_REF(on_movable_entered_occupied_turf),
+	)
+	AddElement(/datum/element/connect_loc, loc_connections)
 
 /obj/item/stack/proc/get_main_recipes()
 	return list()//empty list
@@ -127,18 +126,14 @@
 	if(singular_name)
 		if (latin)
 			if(get_amount()>1)
-				to_chat(user, "There are [get_amount()] [singular_name]i in the stack.")
+				. += "There are [get_amount()] [singular_name]i in the stack."
 			else
-				to_chat(user, "There is [get_amount()] [singular_name]us in the stack.")
+				. += "There is [get_amount()] [singular_name]us in the stack."
 		else
 			if(get_amount()>1)
-				to_chat(user, "There are [get_amount()] [singular_name]\s in the stack.")
+				. += "There are [get_amount()] [singular_name]\s in the stack."
 			else
-				to_chat(user, "There is [get_amount()] [singular_name] in the stack.")
-		if(get_amount()>1)
-			. += "There are [get_amount()] [singular_name]\s in the stack."
-		else
-			. += "There is [get_amount()] [singular_name] in the stack."
+				. += "There is [get_amount()] [singular_name] in the stack."
 	else if(get_amount()>1)
 		. += "There are [get_amount()] in the stack."
 	else
@@ -236,7 +231,7 @@
 			return
 		if (R.time)
 			var/adjusted_time = 0
-			usr.visible_message(span_notice("[usr] starts building [R.title]."), span_notice("You start building [R.title]..."))
+			usr.visible_message("<span class='notice'>[usr] starts building [R.title].</span>", "<span class='notice'>You start building [R.title]...</span>")
 			if(HAS_TRAIT(usr, R.trait_booster))
 				adjusted_time = (R.time * R.trait_modifier)
 			else
@@ -248,11 +243,7 @@
 
 		var/obj/O
 		if(R.max_res_amount > 1) //Is it a stack?
-			if(R.is_stack)
-				O = new R.result_type(usr.drop_location(), R.res_amount * multiplier)
-			else
-				for(var/i in 1 to multiplier)
-					O = new R.result_type(get_turf(usr))
+			O = new R.result_type(usr.drop_location(), R.res_amount * multiplier)
 		else if(ispath(R.result_type, /turf))
 			var/turf/T = usr.drop_location()
 			if(!isturf(T))
@@ -301,22 +292,28 @@
 /obj/item/stack/proc/building_checks(datum/stack_recipe/R, multiplier)
 	if (get_amount() < R.req_amount*multiplier)
 		if (R.req_amount*multiplier>1)
-			to_chat(usr, span_warning("You haven't got enough [src] to build \the [R.req_amount*multiplier] [R.title]\s!"))
+			to_chat(usr, "<span class='warning'>You haven't got enough [src] to build \the [R.req_amount*multiplier] [R.title]\s!</span>")
 		else
-			to_chat(usr, span_warning("You haven't got enough [src] to build \the [R.title]!"))
+			to_chat(usr, "<span class='warning'>You haven't got enough [src] to build \the [R.title]!</span>")
 		return FALSE
 	var/turf/T = get_turf(usr)
 
 	var/obj/D = R.result_type
 	if(R.window_checks && !valid_window_location(T, initial(D.dir) == FULLTILE_WINDOW_DIR ? FULLTILE_WINDOW_DIR : usr.dir))
-		to_chat(usr, span_warning("The [R.title] won't fit here!"))
+		to_chat(usr, "<span class='warning'>The [R.title] won't fit here!</span>")
 		return FALSE
 	if(R.one_per_turf && (locate(R.result_type) in T))
-		to_chat(usr, span_warning("There is another [R.title] here!"))
+		to_chat(usr, "<span class='warning'>There is another [R.title] here!</span>")
 		return FALSE
+	if(R.check_for.len)
+		for(var/check in R.check_for)
+			var/obj/found = locate(check) in T
+			if(found)
+				to_chat(usr, "<span class='warning'>There is a [found.name] here!</span>")
+				return FALSE
 	if(R.on_floor)
 		if(!isfloorturf(T) && !isgroundturf(T))
-			to_chat(usr, span_warning("\The [R.title] must be constructed on the floor!"))
+			to_chat(usr, "<span class='warning'>\The [R.title] must be constructed on the floor!</span>")
 			return FALSE
 		for(var/obj/AM in T)
 			if(istype(AM,/obj/structure/grille))
@@ -328,7 +325,7 @@
 				if(!W.fulltile)
 					continue
 			if(AM.density)
-				to_chat(usr, span_warning("Theres a [AM.name] here. You cant make a [R.title] here!"))
+				to_chat(usr, "<span class='warning'>Theres a [AM.name] here. You cant make a [R.title] here!</span>")
 				return FALSE
 	if(R.placement_checks)
 		switch(R.placement_checks)
@@ -337,23 +334,23 @@
 				for(var/direction in GLOB.cardinals)
 					step = get_step(T, direction)
 					if(locate(R.result_type) in step)
-						to_chat(usr, span_warning("\The [R.title] must not be built directly adjacent to another!"))
+						to_chat(usr, "<span class='warning'>\The [R.title] must not be built directly adjacent to another!</span>")
 						return FALSE
 			if(STACK_CHECK_ADJACENT)
 				if(locate(R.result_type) in range(1, T))
-					to_chat(usr, span_warning("\The [R.title] must be constructed at least one tile away from others of its type!"))
+					to_chat(usr, "<span class='warning'>\The [R.title] must be constructed at least one tile away from others of its type!</span>")
 					return FALSE
 	return TRUE
 
 /obj/item/stack/use(used, transfer = FALSE, check = TRUE) // return 0 = borked; return 1 = had enough
-	if(check && zero_amount())
+	if(check && is_zero_amount(delete_if_zero = TRUE))
 		return FALSE
 	if (is_cyborg)
 		return source.use_charge(used * cost)
 	if (amount < used)
 		return FALSE
 	amount -= used
-	if(check && zero_amount())
+	if(check && is_zero_amount(delete_if_zero = TRUE))
 		return TRUE
 	if(length(mats_per_unit))
 		var/temp_materials = custom_materials.Copy()
@@ -368,23 +365,30 @@
 	if(get_amount() < amount)
 		if(singular_name)
 			if(amount > 1)
-				to_chat(user, span_warning("You need at least [amount] [singular_name]\s to do this!"))
+				to_chat(user, "<span class='warning'>You need at least [amount] [singular_name]\s to do this!</span>")
 			else
-				to_chat(user, span_warning("You need at least [amount] [singular_name] to do this!"))
+				to_chat(user, "<span class='warning'>You need at least [amount] [singular_name] to do this!</span>")
 		else
-			to_chat(user, span_warning("You need at least [amount] to do this!"))
+			to_chat(user, "<span class='warning'>You need at least [amount] to do this!</span>")
 
 		return FALSE
 
 	return TRUE
 
-/obj/item/stack/proc/zero_amount()
+/**
+ * Returns TRUE if the item stack is the equivalent of a 0 amount item.
+ *
+ * Also deletes the item if delete_if_zero is TRUE and the stack does not have
+ * is_cyborg set to true.
+ */
+/obj/item/stack/proc/is_zero_amount(delete_if_zero = TRUE)
 	if(is_cyborg)
 		return source.energy < cost
 	if(amount < 1)
-		qdel(src)
-		return 1
-	return 0
+		if(delete_if_zero)
+			qdel(src)
+		return TRUE
+	return FALSE
 
 /obj/item/stack/proc/add(amount)
 	if (is_cyborg)
@@ -406,9 +410,7 @@
  * - [inhand][boolean]: Whether or not the stack to check should act like it's in a mob's hand.
  */
 /obj/item/stack/proc/can_merge(obj/item/stack/check, inhand = FALSE)
-	if(!istype(check))
-		return FALSE
-	if(check.merge_type != merge_type)
+	if(!istype(check, merge_type))
 		return FALSE
 	if(mats_per_unit ~! check.mats_per_unit) // ~! in case of lists this operator checks only keys, but not values
 		return FALSE
@@ -452,9 +454,14 @@
 		target_stack.mats_per_unit = temp_mats_list
 	return transfer
 
+/**
+ * Merges as much of src into target_stack as possible. If present, the limit arg overrides target_stack.max_amount for transfer.
+ *
+ * This proc deletes src if the remaining amount after the transfer is 0.
+ */
 /obj/item/stack/proc/merge(obj/item/stack/target_stack, limit)
 	. = merge_without_del(target_stack, limit)
-	zero_amount()
+	is_zero_amount(delete_if_zero = TRUE)
 
 /// Signal handler for connect_loc element. Called when a movable enters the turf we're currently occupying. Merges if possible.
 /obj/item/stack/proc/on_movable_entered_occupied_turf(datum/source, atom/movable/arrived)
@@ -465,18 +472,18 @@
 		return
 
 	if(!arrived.throwing && can_merge(arrived))
-		INVOKE_ASYNC(src, PROC_REF(merge), arrived)
+		merge(arrived)
 
-/obj/item/stack/hitby(atom/movable/AM, skipcatch, hitpush, blocked, datum/thrownthing/throwingdatum)
-	if(can_merge(AM, TRUE))
-		merge(AM)
+/obj/item/stack/hitby(atom/movable/hitting, skipcatch, hitpush, blocked, datum/thrownthing/throwingdatum)
+	if(can_merge(hitting, inhand = TRUE))
+		merge(hitting)
 	. = ..()
 
 /obj/item/stack/on_attack_hand(mob/user, act_intent = user.a_intent, unarmed_attack_flags)
 	if(user.get_inactive_held_item() == src)
-		if(zero_amount())
+		if(is_zero_amount(delete_if_zero = TRUE))
 			return
-		return change_stack(user,1)
+		return split_stack(user,1)
 	else
 		. = ..()
 
@@ -487,7 +494,7 @@
 	if(is_cyborg)
 		return
 	else
-		if(zero_amount())
+		if(is_zero_amount(delete_if_zero = TRUE))
 			return
 		//get amount from user
 		var/max = get_amount()
@@ -497,11 +504,17 @@
 		if(stackmaterial == null || stackmaterial <= 0 || !user.canUseTopic(src, BE_CLOSE, ismonkey(user)))
 			return TRUE
 		else
-			change_stack(user, stackmaterial)
-			to_chat(user, span_notice("You take [stackmaterial] sheets out of the stack"))
+			split_stack(user, stackmaterial)
+			to_chat(user, "<span class='notice'>You take [stackmaterial] sheets out of the stack</span>")
 		return TRUE
 
-/obj/item/stack/proc/change_stack(mob/user, amount)
+/** Splits the stack into two stacks.
+ *
+ * Arguments:
+ * - [user][/mob]: The mob splitting the stack.
+ * - amount: The number of units to split from this stack.
+ */
+/obj/item/stack/proc/split_stack(mob/user, amount)
 	if(!use(amount, TRUE, FALSE))
 		return FALSE
 	var/obj/item/stack/F = new type(user? user : drop_location(), amount, FALSE)
@@ -512,13 +525,14 @@
 			F.forceMove(user.drop_location())
 		add_fingerprint(user)
 		F.add_fingerprint(user)
-	zero_amount()
+
+	is_zero_amount(delete_if_zero = TRUE)
 
 /obj/item/stack/attackby(obj/item/W, mob/user, params)
-	if(can_merge(W, TRUE))
+	if(can_merge(W, inhand = TRUE))
 		var/obj/item/stack/S = W
 		if(merge(S))
-			to_chat(user, span_notice("Your [S.name] stack now contains [S.get_amount()] [S.singular_name]\s."))
+			to_chat(user, "<span class='notice'>Your [S.name] stack now contains [S.get_amount()] [S.singular_name]\s.</span>")
 	else
 		. = ..()
 
@@ -553,10 +567,9 @@
 	var/applies_mats = FALSE
 	var/trait_booster = null
 	var/trait_modifier = 1
-	/// Is the resulting thing made from this a stack? if so, multi-crafting will make a stack with multiple 'uses' of it. if not, multi-crafting makes several separate items of it
-	var/is_stack = TRUE
+	var/list/check_for = list()
 
-/datum/stack_recipe/New(title, result_type, req_amount = 1, res_amount = 1, max_res_amount = 1,time = 0, one_per_turf = FALSE, on_floor = FALSE, window_checks = FALSE, placement_checks = FALSE, applies_mats = FALSE, trait_booster = null, trait_modifier = 1, is_stack = TRUE)
+/datum/stack_recipe/New(title, result_type, req_amount = 1, res_amount = 1, max_res_amount = 1,time = 0, one_per_turf = FALSE, on_floor = FALSE, window_checks = FALSE, placement_checks = FALSE, applies_mats = FALSE, trait_booster = null, trait_modifier = 1, list/check_for = list())
 
 
 	src.title = title
@@ -572,7 +585,7 @@
 	src.applies_mats = applies_mats
 	src.trait_booster = trait_booster
 	src.trait_modifier = trait_modifier
-	src.is_stack = is_stack
+	src.check_for = check_for
 /*
  * Recipe list datum
  */
